% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rme-package.R
\docType{package}
\name{rme-package}
\alias{rme}
\alias{rme-package}
\title{rme: Regression Models for Epidemiology}
\description{
Lecture notes for Epi 204 @ UC Davis
}
\author{
\strong{Maintainer}: Douglas Ezra Morrison \email{demorrison@ucdavis.edu} (\href{https://orcid.org/0000-0002-7195-830X}{ORCID}) (@demstats1)

}
\keyword{internal}
